Migrate RefetchHost
