+++ server/service/endpoint_hosts.go
@@ -172,28 +172,3 @@ func makeAddHostsToTeamByFilterEndpoint(svc fleet.Service) endpoint.Endpoint {
 		return addHostsToTeamByFilterResponse{}, err
 	}
 }
-
-////////////////////////////////////////////////////////////////////////////////
-// Refetch Host
-////////////////////////////////////////////////////////////////////////////////
-
-type refetchHostRequest struct {
-	ID uint `json:"id"`
-}
-
-type refetchHostResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r refetchHostResponse) error() error { return r.Err }
-
-func makeRefetchHostEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(refetchHostRequest)
-		err := svc.RefetchHost(ctx, req.ID)
-		if err != nil {
-			return refetchHostResponse{Err: err}, nil
-		}
-		return refetchHostResponse{}, nil
-	}
-}
