+++ server/service/transport_hosts.go
@@ -22,14 +22,6 @@ func decodeDeleteHostRequest(ctx context.Context, r *http.Request) (interface{},
 	return deleteHostRequest{ID: id}, nil
 }
 
-func decodeRefetchHostRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	return refetchHostRequest{ID: id}, nil
-}
-
 func decodeAddHostsToTeamRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	var req addHostsToTeamRequest
